requirements.txt
Flask==3.1.2
flask-cors==6.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.11.3
python-dotenv==1.0.0
gunicorn==21.2.0